+++ src/MsSql/__tests__/password-validation.spec.ts
@@ -3,11 +3,10 @@ import { MsSql } from "../MsSql";
 import { MsSqlPwdValidityEnum, msSqlPwdValidityEnumMessage } from "../MsSqlPwdValidityEnum";
 
 describe("Password validation", () => {
+	const valid = msSqlPwdValidityEnumMessage[MsSqlPwdValidityEnum.Valid];
+
 	test("default password is ok", () => {
-		expect(MsSql.isPasswordValid(MsSql.defaultPassword)).toEqual([
-			true,
-			msSqlPwdValidityEnumMessage[MsSqlPwdValidityEnum.Valid],
-		]);
+		expect(MsSql.isPasswordValid(MsSql.defaultPassword)).toEqual([true, valid]);
 	});
 
 	test("empty password", () => {
@@ -15,6 +14,7 @@ describe("Password validation", () => {
 	});
 
 	test("password too short", () => {
+		expect(MsSql.isPasswordValid("Password12")).toEqual([true, valid]);
 		expect(MsSql.isPasswordValid("Password1")).toEqual([
 			false,
 			msSqlPwdValidityEnumMessage[MsSqlPwdValidityEnum.TooShort],
@@ -22,11 +22,17 @@ describe("Password validation", () => {
 	});
 
 	test("password too simple", () => {
+		expect(MsSql.isPasswordValid("PASSWORD1!")).toEqual([true, valid]);
+		expect(MsSql.isPasswordValid("password1!")).toEqual([true, valid]);
 		expect(MsSql.isPasswordValid("password12")).toEqual([
 			false,
 			msSqlPwdValidityEnumMessage[MsSqlPwdValidityEnum.TooSimple],
 		]);
 		expect(MsSql.isPasswordValid("PASSWORD12")).toEqual([
+			false,
+			msSqlPwdValidityEnumMessage[MsSqlPwdValidityEnum.TooSimple],
+		]);
+		expect(MsSql.isPasswordValid("PASSWORD!!")).toEqual([
 			false,
 			msSqlPwdValidityEnumMessage[MsSqlPwdValidityEnum.TooSimple],
 		]);